requirements.txt
aiodns
cs50
discord.py
matplotlib
python-dotenv
audioop-lts